requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
tzlocal==5.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
